pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dexter"
version = "0.1.0"
description = "AI Assistant"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "David Dominguez", email = "your.email@example.com"}
]
keywords = ["ai", "assistant", "robotics", "speech-to-text", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aioice>=0.10.1",
    "aiortc>=1.13.0",
    "aiosignal>=1.4.0",
    "altair>=4.2.2",
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "attmap>=0.13.2",
    "attrs>=25.3.0",
    "audio-recorder-streamlit>=0.0.10",
    "audioop-lts>=0.2.2",
    "audioread>=3.0.1",
    "av>=14.4.0",
    "azure-cognitiveservices-speech>=1.46.0",
    "azure-core>=1.35.1",
    "beautifulsoup4>=4.13.5",
    "blinker>=1.9.0",
    "cachetools>=5.5.2",
    "certifi>=2025.8.3",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.3",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "coloredlogs>=15.0.1",
    "contourpy>=1.3.3",
    "cryptography>=45.0.7",
    "ctranslate2>=4.6.0",
    "cycler>=0.12.1",
    "ddgs>=9.5.5",
    "decorator>=5.2.1",
    "defusedxml>=0.7.1",
    "distro>=1.9.0",
    "dnspython>=2.8.0",
    "eido>=0.2.4",
    "elevenlabs>=2.15.0",
    "emoji>=2.14.1",
    "entrypoints>=0.4",
    "enum34>=1.1.10",
    "fastapi>=0.116.1",
    "faster-whisper>=1.1.1",
    "fastuuid>=0.12.0",
    "filelock>=3.19.1",
    "flatbuffers>=25.2.10",
    "fonttools>=4.59.2",
    "frozenlist>=1.7.0",
    "fsspec>=2025.9.0",
    "gitdb>=4.0.12",
    "GitPython>=3.1.45",
    "google>=3.0.0",
    "google-api-core>=2.25.1",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-crc32c>=1.7.1",
    "google-genai>=1.36.0",
    "googleapis-common-protos>=1.70.0",
    "h11>=0.16.0",
    "halo>=0.0.31",
    "helium>=5.1.1",
    "hf-xet>=1.1.10",
    "httpcore>=1.0.9",
    "httplib2>=0.31.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.34.4",
    "humanfriendly>=10.0",
    "idna>=3.10",
    "ifaddr>=0.2.0",
    "importlib_metadata>=8.7.0",
    "Jinja2>=3.1.6",
    "jiter>=0.10.0",
    "joblib>=1.5.2",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "kiwisolver>=1.4.9",
    "lazy_loader>=0.4",
    "librosa>=0.11.0",
    "litellm>=1.77.0",
    "llvmlite>=0.44.0",
    "log-symbols>=0.0.14",
    "logmuse>=0.2.8",
    "lxml>=6.0.1",
    "markdown-it-py>=4.0.0",
    "markdownify>=1.2.0",
    "MarkupSafe>=3.0.2",
    "matplotlib>=3.10.6",
    "mdurl>=0.1.2",
    "mpmath>=1.3.0",
    "msgpack>=1.1.1",
    "multidict>=6.6.4",
    "narwhals>=2.5.0",
    "networkx>=3.5",
    "nltk>=3.9.1",
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "onnxruntime>=1.22.1",
    "openai>=1.107.2",
    "opencv-python>=4.12.0.88",
    "openwakeword>=0.6.0",
    "outcome>=1.3.0.post0",
    "oyaml>=1.0",
    "packaging>=25.0",
    "pandas>=2.3.2",
    "pephubclient>=0.4.5",
    "peppy>=0.40.7",
    "pillow>=11.3.0",
    "piper-tts>=1.3.0",
    "pipestat>=0.12.1",
    "platformdirs>=4.4.0",
    "plotly>=6.3.0",
    "pooch>=1.8.2",
    "primp>=0.15.0",
    "propcache>=0.3.2",
    "proto-plus>=1.26.1",
    "protobuf>=6.32.1",
    "psutil>=7.0.0",
    "pvporcupine>=1.9.5",
    "pyarrow>=21.0.0",
    "pyasn1>=0.6.1",
    "pyasn1_modules>=0.4.2",
    "PyAudio>=0.2.14",
    "pycparser>=2.23",
    "pydantic>=2.11.9",
    "pydantic_core>=2.33.2",
    "pydeck>=0.9.1",
    "pyee>=13.0.0",
    "Pygments>=2.19.2",
    "pylibsrtp>=0.12.0",
    "pyobjc>=11.1",
    "pyobjc-core>=11.1",
    "pyOpenSSL>=25.1.0",
    "pyparsing>=3.2.4",
    "pypdf>=6.0.0",
    "PySocks>=1.7.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyttsx3>=2.99",
    "pytz>=2025.2",
    "PyYAML>=6.0.2",
    "realtimestt>=0.3.104",
    "referencing>=0.36.2",
    "regex>=2025.9.1",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "rpds-py>=0.27.1",
    "rsa>=4.9.1",
    "safetensors>=0.6.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.2",
    "selenium>=4.35.0",
    "setuptools>=80.9.0",
    "shellingham>=1.5.4",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "smolagents>=1.21.3",
    "sniffio>=1.3.1",
    "sortedcontainers>=2.4.0",
    "soundfile>=0.13.1",
    "soupsieve>=2.8",
    "soxr>=1.0.0",
    "spinners>=0.0.24",
    "standard-aifc>=3.13.0",
    "standard-chunk>=3.13.0",
    "standard-sunau>=3.13.0",
    "stanza>=1.10.1",
    "starlette>=0.47.3",
    "stream2sentence>=0.3.1",
    "streamlit>=1.49.1",
    "streamlit-webrtc>=0.63.4",
    "sympy>=1.14.0",
    "tenacity>=9.1.2",
    "termcolor>=3.1.0",
    "threadpoolctl>=3.6.0",
    "tiktoken>=0.11.0",
    "tokenizers>=0.22.0",
    "toml>=0.10.2",
    "toolz>=1.0.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "tornado>=6.5.2",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "trio>=0.30.0",
    "trio-websocket>=0.12.2",
    "typer>=0.17.4",
    "typing-inspection>=0.4.1",
    "typing_extensions>=4.14.1",
    "tzdata>=2025.2",
    "ubiquerg>=0.8.1",
    "uritemplate>=4.2.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
    "webrtcvad-wheels>=2.0.14",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
    "wsproto>=1.2.0",
    "yacman>=0.9.3",
    "yarl>=1.20.1",
    "youtube-transcript-api>=1.2.2",
    "zipp>=3.23.0",
]

[project.scripts]
dexter-server = "scripts.server:main"

[project.urls]
Homepage = "https://github.com/Davegdd/DeXteR"
Repository = "https://github.com/Davegdd/DeXteR"
Issues = "https://github.com/Davegdd/DeXteR/issues"

[tool.setuptools.packages.find]
include = ["dexter*"]
exclude = ["tests*"]
